[Bug][#143851][#148508] Create Request Woman Only

+++ src/wsm/wsm/model/LeaveTypeModel.swift
@@ -22,7 +22,7 @@ class LeaveTypeModel: BaseModel {
     var siblingId: String?
     var companyId: String?
     var maxLeaveDuration: Any?
-    var maxPerOne: Int?
+    var maxPerOne: Float?
     var unlimit: Bool?
     var isSuccessive: Bool?
     var numberExpireCompensationDate: Int?
